csrf protection to subject delete related #609

+++ src/Ojs/JournalBundle/Controller/SubjectController.php
@@ -10,6 +10,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Ojs\Common\Controller\OjsController as Controller;
 use Ojs\JournalBundle\Entity\Subject;
 use Ojs\JournalBundle\Form\SubjectType;
+use Symfony\Component\Security\Csrf\Exception\TokenNotFoundException;
 
 /**
  * Subject controller.
@@ -27,8 +28,11 @@ class SubjectController extends Controller
         $source = new Entity("OjsJournalBundle:Subject");
         $grid = $this->get('grid')->setSource($source);
         $actionColumn = new ActionsColumn("actions", 'actions');
+
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
         $rowAction[] = ActionHelper::showAction('subject_show', 'id');
         $rowAction[] = ActionHelper::editAction('subject_edit', 'id');
+        $rowAction[] = ActionHelper::showAction('subject_delete', 'id');
 
         $actionColumn->setRowActions($rowAction);
         $grid->addColumn($actionColumn);
@@ -184,16 +188,22 @@ class SubjectController extends Controller
     }
 
     /**
-     * Deletes a Subject entity.
-     *
+     * @param Request $request
      * @param $id
      * @return RedirectResponse
+     * @throws TokenNotFoundException
      */
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('OjsJournalBundle:Subject')->find($id);
         $this->throw404IfNotFound($entity);
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('subject'.$id);
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
+
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
